Clean up webapp url helpers.
